# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server celler server.",
        "title": "Swagger Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/medical-records": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List Medical Records",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MedicalRecords"
                ],
                "summary": "List Medical Records",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User ID",
                        "name": "user_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Record Type",
                        "name": "record_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Record Date",
                        "name": "record_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Doctor ID",
                        "name": "doctor_id",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_sync.GetAllMedicalRecordsRes"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new Medical Record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MedicalRecords"
                ],
                "summary": "Create a new Medical Record",
                "parameters": [
                    {
                        "description": "body for MedicalRecordCreate",
                        "name": "medicalrecord",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/health_sync.MedicalRecordCreate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/medical-records/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get Medical Record with the given details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MedicalRecords"
                ],
                "summary": "Get Medical Record with the given details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medical Record ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/health_sync.MedicalRecordRes"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update Medical Record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MedicalRecords"
                ],
                "summary": "Update Medical Record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medical Record ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "body for Medical Record Update Body",
                        "name": "Athlete",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/health_sync.MedicalRecordUpdateBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete Medical Record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MedicalRecords"
                ],
                "summary": "Delete Medical Record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Medical Record ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Record deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "error\": \"error message",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "health_sync.GetAllMedicalRecordsRes": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "medical_records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/health_sync.MedicalRecordRes"
                    }
                }
            }
        },
...
        "health_sync.MedicalRecordCreate": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "string"
                },
                "record_date": {
                    "type": "string"
                },
                "record_type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "health_sync.MedicalRecordRes": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "doctor_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "record_date": {
                    "type": "string"
                },
                "record_type": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "health_sync.MedicalRecordUpdateBody": {
            "type": "object",
            "properties": {
                "attachments": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "record_date": {
                    "type": "string"
                },
                "record_type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Description for what is this security definition being used",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}